v4/requirements.txt
# Core deep learning
torch>=2.1
transformers>=4.40
accelerate>=0.27

# ML and stats
scikit-learn>=1.3
scipy>=1.11
numpy>=1.24
pandas>=2.0
pyarrow>=14           # parquet I/O

# Imaging
opencv-python-headless>=4.8
Pillow>=10

# Notebooks + conversion
nbformat>=5.9
jupyter>=1.0
tqdm>=4.65

# Visualisation (exp 05)
matplotlib>=3.7
umap-learn>=0.5       # optional; exp 05 falls back to PCA if absent

# Optional: jupytext round-trip if you prefer editing the .py sources
jupytext>=1.15
